Former-commit-id: ac37a389635153c54cc5082f2acebaad9fce9fc3

+++ server/app/classes/game/coop-game.ts
@@ -31,11 +31,11 @@ export class CoopGame extends Game {
     private endDate: number;
 
     constructor(
-        lobby: CoopLobby, 
-        socketService: SocketService, 
-        private drawingsService: DrawingsService, 
-        private statsService: StatsService, 
-        private userService: UserService, 
+        lobby: CoopLobby,
+        socketService: SocketService,
+        private drawingsService: DrawingsService,
+        private statsService: StatsService,
+        private userService: UserService,
         private chatManagerService: ChatManagerService) {
         super(<Lobby>lobby, socketService);
         for (let player of lobby.getPlayers()) {
@@ -95,7 +95,7 @@ export class CoopGame extends Game {
                 this.vPlayer.sayWrongGuess();
                 this.setupNextDrawing();
             }
-            else{
+            else {
                 this.vPlayer.sayWrongTry();
             }
         }
@@ -130,16 +130,23 @@ export class CoopGame extends Game {
     delay = () => new Promise(res => setTimeout(res, 500));
 
     private endGame(): void {
-        this.endDate = new Date().getTime();
-        clearInterval(this.gameTimerInterval);
-        clearInterval(this.drawingTimerInterval);
-        this.guessesLeft = 0;
-        this.vPlayer.stopDrawing();
-        this.vPlayer.sayEndCoopGame(this.score);
-        this.socketService.getSocket().to(this.id).emit('endGame', { "finalScore": this.score, "virtualPlayerDrawings": this.pastVirtualDrawings, "virtualPlayerIds": this.pastVirtualDrawingsId });
-        this.socketService.getSocket().to(this.id).emit('message', { "user": { username: "System" }, "text": "La partie est maintenant terminée!", "timestamp": 0, "textColor": "#2065d4", chatId: this.id });
-        this.statsService.updateStats(this.gameName, this.gameType, this.getPlayers(), [this.score], this.startDate, this.endDate);
-        this.chatManagerService.deleteChat(this.id);
+        if(this.isGameEnded){
+            return;
+        }
+        else{
+            this.isGameEnded = true;
+            this.gameEnded.next(true);
+            this.endDate = new Date().getTime();
+            clearInterval(this.gameTimerInterval);
+            clearInterval(this.drawingTimerInterval);
+            this.guessesLeft = 0;
+            this.vPlayer.stopDrawing();
+            this.vPlayer.sayEndCoopGame(this.score);
+            this.socketService.getSocket().to(this.id).emit('endGame', { "finalScore": this.score, "virtualPlayerDrawings": this.pastVirtualDrawings, "virtualPlayerIds": this.pastVirtualDrawingsId });
+            this.socketService.getSocket().to(this.id).emit('message', { "user": { username: "System", avatar: -1 }, "text": "La partie est maintenant terminée!", "timestamp": 0, "textColor": "#2065d4", chatId: this.id });
+            this.statsService.updateStats(this.gameName, this.gameType, this.getPlayers(), [this.score], this.startDate, this.endDate);
+            this.chatManagerService.deleteChat(this.id);
+        }
     }
 
     getPlayers(): any {
@@ -219,7 +226,7 @@ export class CoopGame extends Game {
 
     disconnectGame(username: string) {
         const message: string = username + " s'est déconnecté";
-        this.socketService.getSocket().to(this.id).emit('message', { "user": { username: "System" }, "text": message, "timestamp": 0, "textColor": "#2065d4", chatId: this.id });
+        this.socketService.getSocket().to(this.id).emit('message', { "user": { username: "System", avatar: -1 }, "text": message, "timestamp": 0, "textColor": "#2065d4", chatId: this.id });
 
     }
