+++ server/app/classes/game/solo-game.ts
@@ -30,11 +30,11 @@ export class SoloGame extends Game {
     private endDate: number;
 
     constructor(
-        lobby: SoloLobby, 
-        socketService: SocketService, 
-        private drawingsService: DrawingsService, 
-        private statsService: StatsService, 
-        private userService: UserService, 
+        lobby: SoloLobby,
+        socketService: SocketService,
+        private drawingsService: DrawingsService,
+        private statsService: StatsService,
+        private userService: UserService,
         private chatManagerService: ChatManagerService) {
         super(<Lobby>lobby, socketService);
         for (const player of lobby.getPlayers()) {
@@ -56,7 +56,6 @@ export class SoloGame extends Game {
         const drawing = await (await this.vPlayer.getNewDrawing(this.difficulty, this.pastVirtualDrawings));
         this.currentDrawingName = drawing.drawingName;
         this.pastVirtualDrawingsId.push(drawing.drawingId);
-        this.pastVirtualDrawings.push(drawing.drawingName);
         this.pastVirtualDrawings.push(this.currentDrawingName);
         this.startGameTransition();
     }
@@ -96,7 +95,7 @@ export class SoloGame extends Game {
                 this.setupNextDrawing();
                 this.vPlayer.sayWrongGuess();
             }
-            else{
+            else {
                 this.vPlayer.sayWrongTry();
             }
         }
@@ -131,15 +130,22 @@ export class SoloGame extends Game {
     delay = () => new Promise(res => setTimeout(res, 500));
 
     private endGame(): void {
-        this.endDate = new Date().getTime();
-        clearInterval(this.gameTimerInterval);
-        this.guessesLeft = 0;
-        this.vPlayer.stopDrawing();
-        this.socketService.getSocket().to(this.id).emit('endGame', { "finalScore": this.score, "virtualPlayerDrawings": this.pastVirtualDrawings, "virtualPlayerIds": this.pastVirtualDrawingsId });
-        this.vPlayer.sayEndSoloGame(this.score);
-        this.socketService.getSocket().to(this.id).emit('message', { "user": { username: "System" }, "text": "La partie est maintenant terminée!", "timestamp": 0, "textColor": "#2065d4", chatId: this.id });
-        this.statsService.updateStats(this.gameName, this.gameType, this.getPlayers(), [this.score], this.startDate, this.endDate);
-        this.chatManagerService.deleteChat(this.id);
+        if(this.isGameEnded){
+            return;
+        }
+        else{
+            this.isGameEnded = true;
+            this.gameEnded.next(true);
+            this.endDate = new Date().getTime();
+            clearInterval(this.gameTimerInterval);
+            this.guessesLeft = 0;
+            this.vPlayer.stopDrawing();
+            this.socketService.getSocket().to(this.id).emit('endGame', { "finalScore": this.score, "virtualPlayerDrawings": this.pastVirtualDrawings, "virtualPlayerIds": this.pastVirtualDrawingsId });
+            this.vPlayer.sayEndSoloGame(this.score);
+            this.socketService.getSocket().to(this.id).emit('message', { "user": { username: "System", avatar: -1 }, "text": "La partie est maintenant terminée!", "timestamp": 0, "textColor": "#2065d4", chatId: this.id });
+            this.statsService.updateStats(this.gameName, this.gameType, this.getPlayers(), [this.score], this.startDate, this.endDate);
+            this.chatManagerService.deleteChat(this.id);
+        }
     }
 
     getPlayers(): any {
